@@ -17,9 +17,9 @@
     "web-animations-js": "2.3.2"
   },
   "peerDependencies": {
-    "react": "^16.8",
-    "react-dom": "^16.8",
-    "react-router-dom": "^5.0.1"
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
+    "react-router-dom": "^6.14.0"
   },
   "devDependencies": {
     "@babel/core": "^7.8.7",
@@ -37,14 +37,14 @@
     "eslint-plugin-jsx-a11y": "^6.2.3",
     "eslint-plugin-react": "^7.19.0",
     "eslint-plugin-react-hooks": "^1.7.0",
-    "react": "^16.8",
-    "react-dom": "^16.8",
-    "react-router-dom": "^5.0.1",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
+    "react-router-dom": "^6.14.0",
     "rollup-plugin-babel": "^4.4.0",
     "rollup-plugin-peer-deps-external": "^2.2.2",
     "webpack": "^4.42.0",
     "webpack-cli": "^3.3.11",
-    "webpack-dev-server": "^3.10.3"
+    "webpack-dev-server": "^4.15.1"
   },
   "files": [
     "dist"
